Began to flesh out the computer AI a bit more. it's not perfect, but its more human like at least
