+++ src/utilities/playerBoardFunctions.js
@@ -1,0 +1,60 @@
+import { copyBoard, checkIfVacant } from './sharedFunctions';
+import { ships } from './data';
+
+export const placeShip = (row, col, board, shipIdx, vertical, dict) => {
+    if (shipIdx < ships.length) {
+        const newBoard = copyBoard(board);
+        const shipLen = ships[shipIdx].length;
+
+        if (!vertical) {
+            const diff = newBoard[0].length - (col + shipLen);
+
+            if (diff < 0) col += diff;
+            if (!checkIfVacant(row, col, shipLen, board, vertical)) return false;
+
+            for (let k = 0; k < shipLen; k++) {
+                dict[ships[shipIdx].name].push([row, col]);
+                newBoard[row][col].status = 'ship';
+                col++;
+            }
+        } else {
+            const diff = newBoard.length - (row + shipLen);
+
+            if (diff < 0) row += diff;
+            if (!checkIfVacant(row, col, shipLen, board, vertical)) return false;
+
+            for (let k = 0; k < shipLen; k++) {
+                newBoard[row][col].status = 'ship';
+                dict[ships[shipIdx].name].push([row, col]);
+                row++;
+            }
+        }
+        return newBoard;
+    }
+};
+
+export const handleHover = (row, col, board, vertical, readyToPlay, shipIdx) => {
+    if (!readyToPlay && shipIdx < ships.length) {
+        const shipLen = ships[shipIdx].length;
+        const newBoard = copyBoard(board);
+
+        if (!vertical) {
+            const diff = 10 - (col + shipLen);
+            if (diff < 0) col += diff;
+
+            for (let k = 0; k < shipLen; k++) {
+                newBoard[row][col].hover = true;
+                col++;
+            }
+        } else {
+            const diff = 10 - (row + shipLen);
+            if (diff < 0) row += diff;
+
+            for (let k = 0; k < shipLen; k++) {
+                newBoard[row][col].hover = true;
+                row++;
+            }
+        }
+        return newBoard;
+    }
+};
